Updated to .net 5.0.102
commit 307c648daf362be22607dfef18dc89d0a530f818
Author: Brian Greco <blgreco72dev@gmail.com>
Date:   Fri Jan 15 16:18:56 2021 -0500

    Updated to .net 5.0.102

commit c7c533c9476afe8daf4b4b9a95afeb45b261119f
Author: Brian Greco <blgreco72@gmail.com>
Date:   Fri Jan 15 16:18:15 2021 -0500

    Update branches-build.yml

commit 2da9e9a5efa21387d255e4b7e0b0b612ee62f7d7
Author: Brian Greco <blgreco72dev@gmail.com>
Date:   Fri Jan 15 12:21:07 2021 -0500

    Corrected Code Suggestions

commit 907555faf6bf0f678a8edeb2e81fc5a9127277cf
Author: Brian Greco <blgreco72dev@gmail.com>
Date:   Fri Jan 15 08:12:51 2021 -0500

    Deleting no longer needed files

commit b8d01c8fa13946018348fa9beffa63f9132ed772
Merge: 75f0c9e7 c9e710a7
Author: Brian Greco <blgreco72dev@gmail.com>
Date:   Fri Jan 15 08:09:13 2021 -0500

    Merge branch 'develop' into current-feature

+++ netfusion/src/Integration/NetFusion.RabbitMq/Publisher/Internal/DefaultPublisherStrategy.cs
@@ -54,7 +54,7 @@ namespace NetFusion.RabbitMQ.Publisher.Internal
                 ContentType = definition.ContentType,
                 AppId = context.BusModule.HostAppId,
                 Timestamp = DateTimeOffset.UtcNow.ToUnixTimeMilliseconds(),
-                DeliveryMode = (byte) (definition.IsPersistent ? 2 : 1)
+                DeliveryMode = definition.IsPersistent ? 2 : 1
             };
             
             string correlationId = message.GetCorrelationId();
